Move `useMedia` to own file
